Split analytics file 1/2 - create and update
conversionAnalytics file had 2 jobs that were impractical and slow.

1. check dates and see if a day has passed since the creation of the last analytics, if yes, create a new one, if no then do nothing.
if tracked data was preset then increment the data and if startdate and enddate were present then return the sum between the dates.

So I optimized and split them into 2 files thats their only purpose is to do a single job

the first split was the check date and create functionality, the old functions would have too many calls to mongo in order to check when to create a doc and when to update.

it can be simply done by creating the doc with a date of midnight today and checking it when a call happens, if its a doc with midnight today then update, if not then create, and that can be done with "upsert"

Also, I split the names and API URLs on the app.js