Split analytics file 2/2 - search between
Second split was to get the sum of the counted analytics data between 2 given dates.

changed "trackedData" to "count" to make it more understandable and removed the startDate and endDate from the Analytics model which is not needed.

The function stayed almost the same except from the removal of userID when grouping

Split on the app.js API URLs for each functionality